requirements.txt
# Core ML and PyTorch
torch==2.6.0+cu118
torchvision==0.21.0+cu118
torchaudio==2.6.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

transformers>=4.35.0
accelerate>=0.24.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.8.0

# NLP and text processing
spacy>=3.7.0
nltk>=3.8.1

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
joblib>=1.3.0
requests>=2.31.0

# Model optimization and memory efficiency
bitsandbytes>=0.41.0

# Evaluation metrics
rouge-score>=0.1.2

# Visualization and logging (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and debugging
ipython>=8.14.0
jupyter>=1.0.0

# File handling
pathlib2>=2.3.7

# Additional dependencies
unicodedata2>=15.0.0
html5lib>=1.1
